implement executable for storjd + storjutil, rename package to storjd
