+++ bin/_config.js
@@ -1,8 +1,11 @@
 'use strict';
 
+const ini = require('ini');
+const { writeFileSync } = require('fs');
+const mkdirp = require('mkdirp');
 const { homedir } = require('os');
 const { join } = require('path');
-const datadir = join(homedir(), '.config/storj-core');
+const datadir = join(homedir(), '.config/storjd');
 
 module.exports = {
 
@@ -47,3 +50,5 @@ module.exports = {
 
 };
 
+mkdirp.sync(datadir);
+writeFileSync(join(datadir, 'config'), ini.stringify(module.exports));
